requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Deprecated==1.2.14
Django==5.1
django-otp==1.5.1
django-phone-verify==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
nexmo==2.5.2
phonenumbers==8.13.43
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
twilio==9.2.3
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
